Add sequence and catOptional implementations
`sequence` takes an array of optional values and returns an optional
array of unwrapped values. If anything in the array is `.None`, the
function returns `.None`. If all of the values are `.Some`, the function
returns the array of unwrapped values wrapped in `.Some`

`catOptionals` is a similar function, but instead of returning an
optional array, it returns a non-optional array of unwrapped values,
stripping out any `.None` values.