Use multiple if let unwrapping
In Swift 1.2 they silently removed the ability to use `?` as a postfix
operator to remove a single level of optionality. Because of this, we
can't use the nested `map` functions that we were using before. But
since they also added the ability to unwrap multiple optionals inside a
single `if let` statement, we're still able to avoid the gross nested
conditionals.